+++ src/my_app/lib/core/presentation/screens/error_screen.dart
@@ -2,54 +2,60 @@ import 'package:dartx/dartx.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:pull_to_refresh_flutter3/pull_to_refresh_flutter3.dart';
 import 'package:very_good_core/app/themes/text_styles.dart';
 import 'package:very_good_core/app/utils/extensions.dart';
-import 'package:very_good_core/app/utils/hooks.dart';
 
 class ErrorScreen extends HookWidget {
   const ErrorScreen({
     super.key,
     required this.onRefresh,
     required this.errorMessage,
-    this.enablePullDown = true,
   });
 
-  final VoidCallback? onRefresh;
+  final Future<void> Function()? onRefresh;
   final String? errorMessage;
-  final bool enablePullDown;
+
+  @override
+  Widget build(BuildContext context) => onRefresh != null
+      ? RefreshIndicator(
+          onRefresh: onRefresh!,
+          child: _ErrorContent(errorMessage: errorMessage),
+        )
+      : _ErrorContent(errorMessage: errorMessage);
+}
+
+class _ErrorContent extends StatelessWidget {
+  const _ErrorContent({
+    required this.errorMessage,
+  });
+
+  final String? errorMessage;
 
   @override
   Widget build(BuildContext context) {
-    final RefreshController refreshController = useRefreshController();
+    final Color textColor = context.colorScheme.onSurface;
 
-    return SmartRefresher(
-      controller: refreshController,
-      header: const ClassicHeader(),
-      enablePullDown: enablePullDown,
-      onRefresh: onRefresh,
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          const Icon(Icons.error, size: 50),
-          Padding(
-            padding: const EdgeInsets.only(top: 12, bottom: 8),
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: <Widget>[
+        Icon(Icons.error, size: 50, color: context.colorScheme.error),
+        Padding(
+          padding: const EdgeInsets.only(top: 12, bottom: 8),
+          child: Text(
+            context.l10n.common_generic_error,
+            style: AppTextStyle.titleMedium.copyWith(color: textColor),
+          ),
+        ),
+        if (kDebugMode && (errorMessage?.isNotNullOrBlank ?? false))
+          FractionallySizedBox(
+            widthFactor: 0.5,
             child: Text(
-              context.l10n.common_generic_error,
-              style: AppTextStyle.titleMedium,
+              errorMessage!,
+              style: AppTextStyle.titleSmall.copyWith(color: textColor),
+              textAlign: TextAlign.center,
             ),
           ),
-          if (kDebugMode && (errorMessage?.isNotNullOrBlank ?? false))
-            FractionallySizedBox(
-              widthFactor: 0.5,
-              child: Text(
-                errorMessage!,
-                style: AppTextStyle.titleSmall,
-                textAlign: TextAlign.center,
-              ),
-            ),
-        ],
-      ),
+      ],
     );
   }
 }