+++ src/my_app/test/widget/features/home/widgets/empty_post_widget_test.dart
@@ -16,7 +16,9 @@ void main() {
         children: <Widget>[
           GoldenTestDeviceScenario(
             name: 'default',
-            builder: () => const MockLocalization(child: EmptyPost()),
+            builder: () => const MockLocalization(
+              child: EmptyPost(),
+            ),
           ),
         ],
       ),