+++ src/my_app/test/widget/features/profile/screens/profile_screen_test.dart
@@ -5,40 +5,91 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:very_good_core/app/constants/enum.dart';
-import 'package:very_good_core/core/domain/bloc/very_good_core/very_good_core_bloc.dart';
+import 'package:very_good_core/core/domain/bloc/app_core/app_core_bloc.dart';
+import 'package:very_good_core/features/auth/domain/bloc/auth/auth_bloc.dart';
 import 'package:very_good_core/features/profile/presentation/screens/profile_screen.dart';
 
 import '../../../../utils/golden_test_device_scenario.dart';
 import '../../../../utils/mock_material_app.dart';
 import '../../../../utils/test_utils.dart';
-import 'profile_screen_test.mocks.dart';
+import '../../../core/screens/main_screen_test.mocks.dart';
 
-@GenerateNiceMocks(<MockSpec<dynamic>>[MockSpec<VeryGoodCoreBloc>()])
+@GenerateNiceMocks(<MockSpec<dynamic>>[
+  MockSpec<AuthBloc>(),
+  MockSpec<AppCoreBloc>(),
+])
 void main() {
-  late MockVeryGoodCoreBloc veryGoodCoreBloc;
+  late MockAuthBloc authBlocInitial;
+  late MockAuthBloc authBlocLoading;
+  late MockAppCoreBloc appCoreBloc;
+  late Map<AppScrollController, ScrollController> scrollControllers;
 
   setUp(() {
-    veryGoodCoreBloc = MockVeryGoodCoreBloc();
+    authBlocInitial = MockAuthBloc();
+    authBlocLoading = MockAuthBloc();
+    appCoreBloc = MockAppCoreBloc();
 
-    when(veryGoodCoreBloc.stream).thenAnswer(
-      (_) => Stream<VeryGoodCoreState>.fromIterable(<VeryGoodCoreState>[
-        VeryGoodCoreState.initial().copyWith(
-          authStatus: AuthStatus.authenticated,
+    scrollControllers = <AppScrollController, ScrollController>{
+      AppScrollController.home: ScrollController(),
+      AppScrollController.profile: ScrollController(),
+    };
+
+    when(appCoreBloc.stream).thenAnswer(
+      (_) => Stream<AppCoreState>.fromIterable(
+        <AppCoreState>[
+          AppCoreState.initial().copyWith(scrollControllers: scrollControllers),
+        ],
+      ),
+    );
+    when(appCoreBloc.state).thenAnswer(
+      (_) =>
+          AppCoreState.initial().copyWith(scrollControllers: scrollControllers),
+    );
+    when(appCoreBloc.getScrollController(any))
+        .thenAnswer((_) => ScrollController());
+
+    when(authBlocInitial.stream).thenAnswer(
+      (_) => Stream<AuthState>.fromIterable(<AuthState>[
+        AuthState.initial().copyWith(
+          status: AuthStatus.authenticated,
           user: mockUser,
           isLoading: false,
         ),
       ]),
     );
-    when(veryGoodCoreBloc.state).thenAnswer(
-      (_) => VeryGoodCoreState.initial().copyWith(
-        authStatus: AuthStatus.authenticated,
+    when(authBlocInitial.state).thenAnswer(
+      (_) => AuthState.initial().copyWith(
+        status: AuthStatus.authenticated,
         user: mockUser,
         isLoading: false,
       ),
     );
+    when(authBlocLoading.stream).thenAnswer(
+      (_) => Stream<AuthState>.fromIterable(<AuthState>[
+        AuthState.initial().copyWith(
+          status: AuthStatus.authenticated,
+          user: mockUser,
+          isLoading: true,
+        ),
+      ]),
+    );
+    when(authBlocLoading.state).thenAnswer(
+      (_) => AuthState.initial().copyWith(
+        status: AuthStatus.authenticated,
+        user: mockUser,
+        isLoading: true,
+      ),
+    );
   });
-  Widget buildProfileScreen() => BlocProvider<VeryGoodCoreBloc>(
-        create: (BuildContext context) => veryGoodCoreBloc,
+  Widget buildProfileScreen(AuthBloc authBloc) => MultiBlocProvider(
+        providers: <BlocProvider<dynamic>>[
+          BlocProvider<AuthBloc>(
+            create: (BuildContext context) => authBloc,
+          ),
+          BlocProvider<AppCoreBloc>(
+            create: (BuildContext context) => appCoreBloc,
+          ),
+        ],
         child: const MockMaterialApp(
           child: Scaffold(
             body: ProfileScreen(),
@@ -57,7 +108,22 @@ void main() {
         children: <Widget>[
           GoldenTestDeviceScenario(
             name: 'default',
-            builder: buildProfileScreen,
+            builder: () => buildProfileScreen(authBlocInitial),
+          ),
+        ],
+      ),
+    );
+    goldenTest(
+      'renders correctly',
+      fileName: 'profile_screen_loading'.goldensVersion,
+      pumpBeforeTest: (WidgetTester tester) async {
+        await tester.pumpAndSettle();
+      },
+      builder: () => GoldenTestGroup(
+        children: <Widget>[
+          GoldenTestDeviceScenario(
+            name: 'default',
+            builder: () => buildProfileScreen(authBlocLoading),
           ),
         ],
       ),